Move LastPointReader and VectorLastPointReader to org.apache.iotdb.db.query.reader.series
